fix(json-qs): always encode strings
Remove the options that allowed specific keys to use percent-encoding for string values. The consequence is that string values are always encoded the same way.

This change prevents a breaking change from happening if you later “widen” the type of a key that was once a simple string and is now capable of receiving more than just strings. It was reasonably unexpected for type widening to be a breaking change, so I've decided it's better to always encode strings the same way.