feat: rework route handler signatures
Before this commit, a `pathParams` object was always passed as the first argument of a route handler, even if a route had no path parameters (or only one). Now, if a route has no path parameters, the pathParams argument is omitted. If a route has one path parameter, its value is passed directly (no object destructuring required). If a route has multiple path parameters, an array of values is passed (array destructuring is recommended here). The client-side signatures have not changed.

Other changes:

- If a route handler doesn't explicitly type its data arguments, they now default to `any` like a normal function would. This means no runtime validation will occur until the arguments are explicitly typed.

- Tuple types in request data are now correctly handled by the code generator.

- Pagination links are only type-checked if you use the `paginate` function (which is recommended).

- The internal `printTypeLiteral` function is now a string generator.